Separate communication code into class hierarchy to handle the two supported communication modes (LFS, S3)
