+++ p4-fusion/lfscommunication/s3comm.cc
@@ -1,0 +1,15 @@
+#include "s3comm.h"
+
+S3Comm::S3Comm(const std::string& serverURL, const std::string& bucket, const std::string& repository, const std::string& username, const std::string& password)
+	: m_ServerURL(serverURL)
+	, m_Bucket(bucket)
+	, m_Repository(repository)
+	, m_Username(username)
+	, m_Password(password)
+{
+}
+
+LFSClient::UploadResult S3Comm::UploadFile(const std::vector<char>& fileContents) const
+{
+	return LFSClient::UploadResult::Error; // Placeholder implementation
+}
